moved loggedInGuard and restrictTo middlewares to a specific middlewares folder
