+++ src/middlewares/restrict-to.js
@@ -1,0 +1,10 @@
+const restrictTo = (...roles) => {
+  return (req, res, next) => {
+    if (!roles.includes(req['user']['role'])) {
+      return res.status(403).json({ message: 'Not allowed to perform this operation.' });
+    }
+    next();
+  };
+};
+
+module.exports = { restrictTo };
