+++ src/routes/users.router.js
@@ -1,7 +1,9 @@
 const express = require('express');
 
-const { signup, login, loggedInGuard, restrictTo } = require('./../controllers/auth.controller');
-const { getAllActiveUsers, disableUser, findUserById, findMyProfile, updateUserProfile } = require('./../controllers/users.controller');
+const { signup, login } = require('./../controllers/auth.controller');
+const { loggedInGuard } = require('./../middlewares/logged-in-guard');
+const { restrictTo } = require('./../middlewares/restrict-to');
+const { getAllActiveUsers, disableUser, findUserById, currentUserProfile, updateUserProfile } = require('./../controllers/users.controller');
 
 const usersRouter = express.Router();
 
@@ -9,7 +11,7 @@ usersRouter.post('/signup', signup);
 
 usersRouter.post('/login', login);
 
-usersRouter.get('/my-profile', loggedInGuard, findMyProfile);
+usersRouter.get('/my-profile', loggedInGuard, currentUserProfile);
 
 usersRouter.patch('/:id', loggedInGuard, updateUserProfile);
